mctpd: Fix native compilation problem
Following error is failing compilation when building natively:
../src/PCIeBinding.cpp:388:49: error: conversion from ‘int’ to ‘uint8_t’ {aka ‘unsigned char’} may change value [-Werror=conversion]

This change adds necessary cast to tell the compiler we know what
we're doing.

+++ mctpd/src/PCIeBinding.cpp
@@ -385,7 +385,7 @@ bool PCIeBinding::handleGetVdmSupport(mctp_eid_t destEid, void* bindingPrivate,
     }
     else
     {
-        resp->vendor_id_set_selector = setIndex + 1;
+        resp->vendor_id_set_selector = static_cast<uint8_t>(setIndex + 1U);
     }
     resp->vendor_id_format = vdmSetDatabase[setIndex].idFormat;
     resp->vendor_id_data = vdmSetDatabase[setIndex].idData;
